src/backend/requirements.txt
flask
flask-cors
paho-mqtt
requests